Combine thread and message renderer
Mostly a bit of code cleanup, side effect feature addition is that we
now support automatic embeds in messages too.
